Module yo::http1::connection

Structs

Connection

Represents a connection with a client or server, managing roles, state, and communication.

InformationalResponse

Represents an informational HTTP response with status code, headers, and reason.

ReceiveBuffer

A buffer for receiving and storing data with efficient extraction operations.

Request

Represents an HTTP request with method, headers, and HTTP version.

Response

Represents an HTTP response with status code, headers, reason, and HTTP version.

Enums

ConnectionState

Represents various states a connection can be in.

Event

Represents different types of events in the system.

Role

Roles for either a client or server.

Sentinel

Sentinel values used to represent certain states in a process.

Constants

DEFAULT_MAX_INCOMPLETE_EVENT_SIZE

The default maximum size for incomplete events, set to 16 KB.

Functions

get_comma_header

Retrieves a comma-separated list of values from the headers for the given key.

set_comma_header

Sets a comma-separated list of values for the given key in the headers.